+++ packages/admin-theme/src/cometAdminOverrides/tableBodyRow.ts
@@ -1,0 +1,9 @@
+import { CometAdminTableBodyRowClassKeys } from "@comet/admin";
+import { StyleRules } from "@material-ui/styles/withStyles";
+
+export const cometAdminTableBodyRowOverrides = (): StyleRules<{}, CometAdminTableBodyRowClassKeys> => ({
+    even: {},
+    odd: {
+        backgroundColor: "#fff",
+    },
+});